+++ src/test/java/pageObgects/homePage/HomePageLocators.java
@@ -1,0 +1,10 @@
+package pageObgects.homePage;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebElement;
+import runner.BaseTest;
+
+public class HomePageLocators extends BaseTest {
+
+    public static WebElement LogOutBtn = getDriver().findElement(By.xpath("//a[@href='/logout']"));
+}
