+++ src/test/java/runner/BaseUtils.java
@@ -4,6 +4,7 @@ import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
+import pageObgects.loginPage.LoginPageLocators;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -72,7 +73,6 @@ public final class BaseUtils {
     static WebDriver createDriver() {
         WebDriver driver = new ChromeDriver(chromeOptions);
         driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
-
         return driver;
     }
 }
