+++ src/test/java/runner/ProjectUtils.java
@@ -2,8 +2,9 @@ package runner;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
 
+import static pageObgects.homePage.HomePageLocators.*;
+import static pageObgects.loginPage.LoginPageLocators.*;
 import static runner.BaseUtils.PREFIX_PROP;
 import static runner.BaseUtils.getProperties;
 
@@ -15,18 +16,12 @@ public final class ProjectUtils {
 
     static void login(WebDriver driver) {
         driver.get(String.format("http://localhost:%s", getProperties().getProperty(PROP_PORT)));
-
-        WebElement name = driver.findElement(By.name("j_username"));
-        name.sendKeys(getProperties().getProperty(PROP_ADMIN_USERNAME));
-
-        WebElement password = driver.findElement(By.name("j_password"));
-        password.sendKeys(getProperties().getProperty(PROP_ADMIN_PAS));
-
-        WebElement SignIn = driver.findElement(By.name("Submit"));
-        SignIn.click();
+        UserNameField.sendKeys(getProperties().getProperty(PROP_ADMIN_USERNAME));
+        PasswordField.sendKeys(getProperties().getProperty(PROP_ADMIN_PAS));
+        SignInBtn.click();
     }
 
-    static void logout(WebDriver driver) {
-        driver.findElement(By.xpath("//a[@href='/logout']")).click();
+    static void logout() {
+        LogOutBtn.click();
     }
 }
